+++ src/components/Home/HomeComponent.module.css
@@ -1,9 +1,5 @@
 
 
-body {
-  margin: 0;
-  overflow: hidden; 
-}
 
 .videoContainer {
   position: relative; /* 상대적 위치 설정 */
