+++ src/containers/home/HomeContainer.tsx
@@ -1,15 +1,15 @@
 /* 데이터 처리: 외부 API로부터 데이터를 가져오거나, Redux 같은 상태 관리 라이브러리를 사용하여 애플리케이션의 상태를 관리
 예시: 페이지, 데이터를 표시하는 리스트 뷰, 사용자 입력을 처리하는 폼 등 */
 
-import containerStyle from "./PromotionContainer.module.css";
+import containerStyle from "./HomeContainer.module.css";
 
-function PromotionContainer() {
+function HomeContainer() {
   return (
     <div className={containerStyle.wrapper}>
     </div>
   );
 }
 
-export default PromotionContainer;
+export default HomeContainer;
