+++ src/routes/Home.tsx
@@ -1,5 +1,16 @@
+import Header from "../components/header/Header";
+import HomeComponent from "../components/Home/HomeComponent";
+import HomeContainer from "../containers/home/HomeContainer";
+
+
 function Home() {
-  return <div></div>;
+  return (
+    <div>
+      <Header />
+      <HomeComponent />
+      <HomeContainer />
+    </div>
+  );
 }
 
 export default Home;
